Add support DTO `Model::class`.
